+++ src/components/dropdowns/user-menu.js
@@ -290,8 +290,8 @@ UserMenu.defaultProps = {
   // Add custom value to compute horizontal position
   spaceAroundGrid: 0,
   // Notifications count on badge.
-  buttonNotifications: 0 // DEPRECATED: do not use default export.
+  buttonNotifications: 0
+}; // DEPRECATED: do not use default export.
 
-};
 var _default = UserMenu;
 exports.default = _default;